chore: Move dashboard content banner tests into their own file
