+++ src/app/(proper_react)/redesign/(authenticated)/user/dashboard/Dashboard.test.tsx
@@ -36,19 +36,12 @@ import Meta, {
   DashboardUsPremiumScanEmptyInProgressUnresolvedBreaches,
   DashboardUsPremiumScanUnresolvedInProgressNoBreaches,
   DashboardUsPremiumScanUnresolvedInProgressUnresolvedBreaches,
-  DashboardInvalidNonPremiumUserScanUnresolvedInProgressResolvedBreaches,
 } from "./Dashboard.stories";
 
 jest.mock("next/navigation", () => ({
   useRouter: jest.fn(),
   usePathname: jest.fn(),
 }));
-
-function getTopBannerContentCta(ctas: HTMLElement[]) {
-  return ctas.filter((element) =>
-    element.parentElement?.parentElement?.classList.contains("explainerContent")
-  )[0];
-}
 
 it("passes the axe accessibility test suite 1", async () => {
   const ComposedDashboard = composeStory(DashboardNonUsNoBreaches, Meta);
@@ -294,384 +287,6 @@ it("passes the axe accessibility test suite 28", async () => {
   );
   const { container } = render(<ComposedDashboard />);
   expect(await axe(container)).toHaveNoViolations();
-});
-
-it("show the correct dashboard top banner content in the story NonUsNoBreaches", () => {
-  const ComposedDashboard = composeStory(DashboardNonUsNoBreaches, Meta);
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Monitor more emails",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story NonUsUnresolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardNonUsUnresolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Let’s fix it",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story NonUsResolvedBreaches", () => {
-  const ComposedDashboard = composeStory(DashboardNonUsResolvedBreaches, Meta);
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("button", {
-    name: "See what’s fixed",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsNoPremiumNoScanNoBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsNoPremiumNoScanNoBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Get first scan free",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsNoPremiumNoScanUnresolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsNoPremiumNoScanUnresolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Get first scan free",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsNoPremiumNoScanResolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsNoPremiumNoScanResolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Get first scan free",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsNoPremiumEmptyScanNoBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsNoPremiumEmptyScanNoBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Get continuous protection",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsNoPremiumEmptyScanUnresolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsNoPremiumEmptyScanUnresolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Let’s keep going",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsNoPremiumEmptyScanResolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsNoPremiumEmptyScanResolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("button", {
-    name: "Get continuous protection",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsNoPremiumUnresolvedScanNoBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsNoPremiumUnresolvedScanNoBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Let’s fix it",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsNoPremiumUnresolvedScanUnresolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsNoPremiumUnresolvedScanUnresolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Let’s keep going",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsNoPremiumUnresolvedScanResolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsNoPremiumUnresolvedScanResolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Let’s fix it",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsNoPremiumResolvedScanNoBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsNoPremiumResolvedScanNoBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("button", {
-    name: "Get continuous protection",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsNoPremiumResolvedScanUnresolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsNoPremiumResolvedScanUnresolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Let’s keep going",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsNoPremiumResolvedScanResolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsNoPremiumResolvedScanResolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("button", {
-    name: "Get continuous protection",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsPremiumEmptyScanNoBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsPremiumEmptyScanNoBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Monitor more emails",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsPremiumEmptyScanUnresolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsPremiumEmptyScanUnresolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Let’s keep going",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsPremiumEmptyScanResolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsPremiumEmptyScanResolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("button", {
-    name: "See what’s fixed",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsPremiumUnresolvedScanNoBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsPremiumUnresolvedScanNoBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Let’s keep going",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsPremiumUnresolvedScanUnresolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsPremiumUnresolvedScanUnresolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Let’s fix it",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsPremiumUnresolvedScanResolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsPremiumUnresolvedScanResolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Let’s keep going",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsPremiumResolvedScanNoBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsPremiumResolvedScanNoBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("button", {
-    name: "See what’s fixed",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsPremiumResolvedScanUnresolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsPremiumResolvedScanUnresolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "Let’s keep going",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsPremiumResolvedScanResolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsPremiumResolvedScanResolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("button", {
-    name: "See what’s fixed",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsPremiumScanEmptyInProgressNoBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsPremiumScanEmptyInProgressNoBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-});
-
-it("show the correct dashboard top banner content in the story UsPremiumScanEmptyInProgressUnresolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsPremiumScanEmptyInProgressUnresolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-});
-
-it("show the correct dashboard top banner content in the story UsPremiumScanUnresolvedInProgressNoBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsPremiumScanUnresolvedInProgressNoBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "See what’s ready now",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story UsPremiumScanUnresolvedInProgressUnresolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardUsPremiumScanUnresolvedInProgressUnresolvedBreaches,
-    Meta
-  );
-  render(<ComposedDashboard />);
-
-  const dashboardTopBannerCta = screen.queryAllByRole("link", {
-    name: "See what’s ready now",
-  });
-  expect(getTopBannerContentCta(dashboardTopBannerCta)).toBeInTheDocument();
-});
-
-it("show the correct dashboard top banner content in the story InvalidNonPremiumUserScanUnresolvedInProgressResolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardInvalidNonPremiumUserScanUnresolvedInProgressResolvedBreaches,
-    Meta
-  );
-
-  const warnLogSpy = jest.spyOn(global.console, "warn").mockImplementation();
-  render(<ComposedDashboard />);
-
-  expect(warnLogSpy).toHaveBeenCalledWith(
-    "No matching condition for dashboard state found."
-  );
-  warnLogSpy.mockReset();
-});
-
-it("show the correct dashboard top banner content in the story DashboardInvalidNonPremiumUserScanUnresolvedInProgressResolvedBreaches", () => {
-  const ComposedDashboard = composeStory(
-    DashboardInvalidNonPremiumUserScanUnresolvedInProgressResolvedBreaches,
-    Meta
-  );
-
-  const warnLogSpy = jest.spyOn(global.console, "warn").mockImplementation();
-  render(<ComposedDashboard />);
-
-  expect(warnLogSpy).toHaveBeenCalledWith(
-    "No matching condition for dashboard state found."
-  );
-  warnLogSpy.mockReset();
 });
 
 it("shows the 'Start a free scan' CTA to free US-based users who haven't performed a scan yet", () => {
